+++ FileManager/Base.lproj/Main.storyboard
@@ -7,10 +7,10 @@
         <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
     </dependencies>
     <scenes>
-        <!--Table View Controller-->
+        <!--Folder View Controller-->
         <scene sceneID="5ZX-NS-5fR">
             <objects>
-                <tableViewController id="ymY-no-j5x" customClass="TableViewController" customModule="FileManager" customModuleProvider="target" sceneMemberID="viewController">
+                <tableViewController id="ymY-no-j5x" customClass="FolderViewController" customModule="FileManager" customModuleProvider="target" sceneMemberID="viewController">
                     <tableView key="view" clipsSubviews="YES" contentMode="scaleToFill" alwaysBounceVertical="YES" dataMode="prototypes" style="plain" separatorStyle="default" rowHeight="-1" estimatedRowHeight="-1" sectionHeaderHeight="-1" estimatedSectionHeaderHeight="-1" sectionFooterHeight="-1" estimatedSectionFooterHeight="-1" id="Jvt-fF-kjZ">
                         <rect key="frame" x="0.0" y="0.0" width="393" height="852"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
@@ -31,7 +31,11 @@
                         </connections>
                     </tableView>
                     <navigationItem key="navigationItem" id="vKH-l1-nCZ">
-                        <barButtonItem key="rightBarButtonItem" title="Add picture" id="gYY-PD-dE9"/>
+                        <barButtonItem key="rightBarButtonItem" title="Add picture" id="gYY-PD-dE9">
+                            <connections>
+                                <action selector="addPicture:" destination="ymY-no-j5x" id="hHr-rm-w9S"/>
+                            </connections>
+                        </barButtonItem>
                     </navigationItem>
                 </tableViewController>
                 <placeholder placeholderIdentifier="IBFirstResponder" id="Ttm-s3-Cr1" userLabel="First Responder" customClass="UIResponder" sceneMemberID="firstResponder"/>
